requirements.txt
pyTelegramBotAPI
requests
aiohttp
python-dotenv
pyncm